Add '# type: ignore' to any `import cupy as cp` statements
If editing the codebase in an environment without CuPy installed (e.g., Mac's) then the editor rightly complains that it cannot find the CuPy installation. In this scenario, we instead rely on the fallback of NumPy by catching the `ImportError`. Therefore, to stop the warning appearing in the editor, we append '# type: ignore' to the line.